simplified the secondary index content (datatype was too verbose) updated the session listing on ghe group analysis page

+++ ui/src/components/groupanalysis.vue
@@ -6,63 +6,76 @@
         </div>
         <iframe :src="host+'/ipython/'+selected.port+'/lab?token='+selected.token" frameBorder="0"/>
     </div>
-    <div v-else-if="ready" class="page-content">
-        <div class="list" style="background-color: white;">
-            <p>
-                Once you complete your subject level processing, you can run group analysis on jupyter hub hosted on brainlife. You can access any group-analysis ready data objects created by running Apps.
-            </p>
-            <p>
-                Please select a group analysis App from the list of available group analysis Apps below,
-                TODO..
-            </p>
 
-            <h4>Existing Notebooks</h4>
-            <div>
-                <div style="display: inline-block; width: 300px; margin-right: 10px; margin-bottom: 10px;" 
-                    v-for="task in tasks.filter(task=>task.status != 'removed')" :key="task._id">
-                    <div class="card ga-card">
-                        <div class="card-body">
-                            <h5 class="card-title" style="min-height: 50px;"><icon name="flask"/> {{task.name}}</h5>
-                            <p class="card-text" style="min-height: 80px;">{{task.desc}}<br> 
-                                <small>{{task.config.container}}</small>
-                            </p>
-                            <b-button variant="primary" size="sm" @click="open(task)" :disabled="task._id == openWhenReady">
-                                <span v-if="task._id == openWhenReady"><icon name="cog" spin/> Opening..</span>
-                                <span v-else><icon name="play"/> Open</span>
-                            </b-button>
-                            <div class="button" style="float: right;" variant="secondary" size="sm" @click="remove(task)"><icon name="trash"/></div>
-                        </div>
-                        <div class="card-footer">
-                            <!--<small class="text-muted"><statustag :status="task.status"/></small><br>-->
-                            <small style="font-size: 70%">{{task.status_msg}} {{task._id}}</small><br>
-                        </div>
-                    </div>
-                </div>
+    <div v-if="!selected && !ready" class="page-content">
+        <p style="padding: 20px">Loading...</p>
+    </div>
+
+    <div v-if="!selected && ready" class="page-content">
+        <!-- instructions -->
+        <p>
+            <b-alert :show="true" variant="secondary">
+                Once you complete your subject level processing, you can run group analysis on jupyter notebook hosted on brainlife.io.
+            </b-alert>
+        </p>
+
+        <div style="padding: 0 20px">
+            <h4>Sessions</h4>
+            <small v-if="sessions.length == 0">Please launch a new session.</small>
+            <div v-for="task in sessions" :key="task._id" class="session">
+                <b-row>
+                    <b-col cols="4">
+                        <small>{{task.name}}</small>
+                        {{task.desc}}
+                        <!--{{task.desc}}<br> -->
+                    </b-col>
+                    <b-col cols="3">
+                        <statusicon :status="task.status"/> <small>{{task._id}}</small><br>
+                        <pre>{{task.status_msg}}</pre>
+                    </b-col>
+                    <b-col cols="2">
+                        Created on <b>{{new Date(task.create_date).toLocaleDateString()}}</b>
+                    </b-col>
+                    <b-col cols="3">
+                        <b-button variant="primary" size="sm" @click="open(task)" :disabled="task._id == openWhenReady">
+                            <span v-if="task._id == openWhenReady"><icon name="cog" spin/> Opening..</span>
+                            <span v-else><!--<icon name="play"/>-->Open</span>
+                        </b-button>
+                        <b-button variant="secondary" size="sm" @click="remove(task)"><!--<icon name="trash"/>-->Remove</b-button>
+                    </b-col>
+                </b-row>
             </div>
-
             <br>
         </div>
 
-        <div class="list">
-            <h4>New Analysis</h4>
-            <div>
-                <div style="display: inline-block; width: 200px; margin-right: 10px; margin-bottom: 10px;" v-for="(app, idx) in apps" :key="idx">
-                    <div class="card ga-card">
-                        <!--<img class="card-img-top" :src="app.img">-->
-                        <div class="card-body">
-                            <b-card-text style="min-height: 60px; font-size: 90%">{{app.desc}}</b-card-text>
-                            <p>
-                                <b-button variant="success" size="sm" @click="launch(app)"><icon name="play"/> Launch</b-button>
-                            </p>
-                        </div>
+        <div class="new">
+            <h4>Launch New Session</h4>
+            <b-form-select v-model="selectedNewApp" @change="selectNewApp" :options="apps" placeholder="Please select a class to launch"/>
+            <!--
+            <div style="display: inline-block; width: 200px; margin-right: 10px; margin-bottom: 10px;" v-for="(app, idx) in apps" :key="idx">
+                <div class="card ga-card">
+                    <div class="card-body">
+                        <b-card-text style="min-height: 60px; font-size: 90%">{{app.desc}}</b-card-text>
+                        <p>
+                            <b-button variant="success" size="sm" @click="launch(app)"><icon name="play"/> Launch</b-button>
+                        </p>
                     </div>
                 </div>
             </div>
+            -->
+            <div v-if="newApp">
+                <p>
+                    <small>{{newApp.container}}:{{newApp.tag}}</small>
+                </p>
+                <p>
+                    <b>Description</b>
+                    <b-form-textarea v-model="newApp.desc"/>
+                </p>
+                <b-button variant="success" size="sm" @click="launchNewApp" style="float: right;"><icon name="play"/> Launch</b-button>
+                <br clear="right">
+            </div>
         </div>
 
-    </div>
-    <div v-else class="page-content">
-        <p style="padding: 20px">Loading...</p>
     </div>
 </div>
 </template>
@@ -80,6 +93,7 @@ export default {
     components: {
         'contact': ()=> import('@/components/contact'),
         'statustag': ()=> import('@/components/statustag'),
+        'statusicon': ()=> import('@/components/statusicon'),
     },
 
     props: {
@@ -95,6 +109,9 @@ export default {
 
             openWhenReady: null,
 
+            selectedNewApp: null, //not really used by needed to show the placeholder
+            newApp: null,
+
             instance: null,
 
             ws: null, //websocket
@@ -102,41 +119,55 @@ export default {
             //will come from db someday
             apps: [
                 {
-                    name: "Brainlife dipy Jupyter Notebook", 
-                    //img: "https://kanoki.org/wp-content/uploads/2017/07/Screen-Shot-2017-07-15-at-04.59.36.png",
-                    desc: "Jupyter Datascience Notebook (lab-2.1.1) with Dipy and Fury",
-                    container: "brainlife/ga-dipy",
-                    tag: "1.0",
+                    text: "Please select a class to start",
+                    value: null,
                 },
                 {
-                    name: "Brainlife Octave(matlab) Jupyter Notebook", 
+                    text: "python/dipy", 
                     //img: "https://kanoki.org/wp-content/uploads/2017/07/Screen-Shot-2017-07-15-at-04.59.36.png",
-                    desc: "Jupyter Datascience Notebook (lab-2.1.1) with Octave",
-                    container: "brainlife/ga-octave",
-                    tag: "1.0",
+                    value: {
+                        desc: "Jupyter Datascience Notebook (lab-2.1.1) with Dipy and Fury",
+                        container: "brainlife/ga-dipy",
+                        tag: "1.0",
+                    },
                 },
                 {
-                    name: "Jupyter Datascience Lab", 
+                    text: "Octave(matlab)", 
                     //img: "https://kanoki.org/wp-content/uploads/2017/07/Screen-Shot-2017-07-15-at-04.59.36.png",
-                    desc: "Jupyter Datascience Notebook (lab-2.1.1)",
-                    container: "jupyter/datascience-notebook", 
-                    tag: "lab-2.1.1", 
+                    value: {
+                        desc: "Jupyter Datascience Notebook (lab-2.1.1) with Octave",
+                        container: "brainlife/ga-octave",
+                        tag: "1.0",
+                    }
                 },
                 {
-                    name: "Jupyter Scipy Notebook", 
+                    text: "jupyter/datascience-notebook", 
+                    //img: "https://kanoki.org/wp-content/uploads/2017/07/Screen-Shot-2017-07-15-at-04.59.36.png",
+                    value: {
+                        desc: "Jupyter Datascience Notebook (lab-2.1.1)",
+                        container: "jupyter/datascience-notebook", 
+                        tag: "lab-2.1.1", 
+                    }
+                },
+                {
+                    text: "jupyter/scipy-notebook", 
                     //img: "https://kanoki.org/wp-content/uploads/2017/07/Screen-Shot-2017-07-15-at-04.59.36.png",
                     //img: "https://www.dataquest.io/wp-content/uploads/2019/01/interface-screenshot.png",
-                    desc: "Jupyter Notebook Scientific Python Stack",
-                    container: "jupyter/scipy-notebook", 
-                    tag: "latest",
+                    value: {
+                        desc: "Jupyter Notebook Scientific Python Stack",
+                        container: "jupyter/scipy-notebook", 
+                        tag: "latest",
+                    }
                 },   
                 {
-                    name: "Jupyter Tensorflow Notebook", 
+                    text: "python/tensorflow", 
                     //img: "https://i.imgur.com/n0PmXQn.gif",
                     //img: "https://kanoki.org/wp-content/uploads/2017/07/Screen-Shot-2017-07-15-at-04.59.36.png",
-                    desc: "Jupyter Notebook Scientific Python Stack w/ Tensorflow",
-                    container: "jupyter/tensorflow-notebook", 
-                    tag: "latest",
+                    value: {
+                        desc: "Jupyter Notebook Scientific Python Stack w/ Tensorflow",
+                        container: "jupyter/tensorflow-notebook", 
+                        tag: "latest",
+                    }
                 },
             ],
 
@@ -203,7 +234,14 @@ export default {
         });
     },
 
+    computed: {
+        sessions() {
+            return this.tasks.filter(task=>task.status != 'removed');
+        }   
+    },
+
     methods: {
+
         find_or_create_instance(cb) {
             //find or create an instance to host all ga tasks
             let key = {
@@ -223,18 +261,26 @@ export default {
             }).catch(cb);
         },
 
-        launch(app) {
+        selectNewApp(app) {
+            this.newApp = Object.assign({}, this.newApp, {
+                desc: app.desc,
+                container: app.container,
+                tag: app.tag,
+            });
+        },
+
+        launchNewApp() {
             this.check_agreements(this.project, err=>{
                 if(err) return this.cb(err);
                 this.$http.post('secondary/launchga', {
                     instance_id: this.instance._id,
 
-                    container: app.container,
+                    container: this.newApp.container,
+                    name: this.newApp.container+":"+this.newApp.tag,
                     app: "soichih/ga-test", //TODO
-                    tag: app.tag,
+                    tag: this.newApp.tag,
 
-                    name: app.name, 
-                    desc: app.desc, //TODO..
+                    desc: this.newApp.desc, 
 
                     //config: {},
                 }).then(res=>{
@@ -246,11 +292,13 @@ export default {
         },
 
         remove(task) {
-            this.$http.delete(Vue.config.amaretti_api+"/task/"+task._id).then(res=>{
-                if(res.status == 200) {
-                    this.$notify("Removal request submitted");
-                }
-            });
+            if(confirm("Do you really want to remove this session?")) {
+                this.$http.delete(Vue.config.amaretti_api+"/task/"+task._id).then(res=>{
+                    if(res.status == 200) {
+                        this.$notify("Removal request submitted");
+                    }
+                });
+            }
         },
 
         open(task) {
@@ -314,22 +362,11 @@ export default {
     top: 95px;
     overflow: auto;
 }
-.list h4 {
-opacity: 0.4;
-text-transform: uppercase;
-font-size: 13pt;
-font-weight: bold;
-margin-bottom: 20px;
-}
-.welcome {
-    margin: 20px;
-}
-.list {
-    padding: 20px;
-}
-.new {
-    padding: 20px;
-    background-color: white;
+h4 {
+    opacity: 0.4;
+    text-transform: uppercase;
+    font-size: 13pt;
+    font-weight: bold;
 }
 iframe {
     position: fixed;
@@ -338,10 +375,6 @@ iframe {
     width: calc(100% - 40px);
     height: calc(100% - 95px);
     transition: left 0.2s;
-}
-.sidewide iframe {
-    left: 180px;
-    width: calc(100% - 180px);
 }
 
 .selected-controller {
@@ -354,24 +387,22 @@ iframe {
     top: 95px;
     z-index: 1;
 }
-.ga-container {
-    padding: 5px;
+
+.session {
+    box-shadow: 1px 1px 2px #0002;
+    padding: 10px;    
     background-color: white;
-    margin-bottom: 1px;
 }
-.ga-container.header {
-    background-color: inherit;
-    opacity: 0.5;
-    text-transform: uppercase;
-    font-weight: bold;
+
+.sidewide iframe {
+    left: 180px;
+    width: calc(100% - 180px);
 }
-.ga-card {
-border-radius: 5px;
-margin-bottom: 10px;
-box-shadow: 2px 2px 4px #0001;
+.new {
+    background-color: white;
+    padding: 20px;
+    box-shadow: 1px 1px 2px #0002;
 }
-.ga-card .card-body {
-padding: 10px;
-}
+
 </style>
